Add code to display the categorization of the triangulation
MonotoneTriangulator now has code in the calculate method to
generate an auxiliary file containing the categorized vertices
called holes-categories.txt. A python file, display.py, takes
this file (and others that are not generated yet) and displays
the triangulation graphically. The code in display.py was written
by Professor Walker White.

The .gitignore file is set to ignore the auxiliary file
holes-categories.txt.

There are new import statements in MonotoneTriangulator.