Add various changes while meeting Walker White on 9/8/2021
Fix bugs in MonotoneTriangulator (specifically the method
monoTriangulate), so that the program now works correctly.

Add code to round numbers to 3 decimal places in graphical
debugging code (in the method calculate in MonotoneTriangulator).

Remove class IntIntMap (from LibGDX) because it seemed like too
much work to get it working correctly in the current time frame,
for too little of a payoff.